fixed the 0-arg asyncSpawns
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@3896 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.lib/src/types.h
@@ -1,7 +1,7 @@
 /*
  * (c) Copyright IBM Corporation 2007
  *
- * $Id: types.h,v 1.6 2007-05-31 11:25:57 ganeshvb Exp $
+ * $Id: types.h,v 1.7 2007-06-16 16:20:36 ganeshvb Exp $
  * This file is part of X10 Runtime System.
  */
 
@@ -13,7 +13,7 @@
 typedef uint32_t place_t;
 typedef uint64_t gas_ref_t;
 typedef int64_t async_arg_t;
-typedef uint8_t async_handler_t;
+typedef int32_t async_handler_t;
 typedef void (*void_func_t) ();
 typedef  struct
 {
